refactor: разделение бизнес-логики и UI логики в AuthService
- Создан SessionManager для управления UI логикой сессии
- Очищен AuthService от UI зависимостей
- Добавлен SessionProvider для глобального доступа к SessionManager
- Обновлена архитектура для разделения ответственности
- Добавлены fallback механизмы для безопасной работы
- Исправлена ошибка _TypeError с SessionManager

+++ lib/shared/models/receipt_model.dart
@@ -50,13 +50,13 @@ class ReceiptProduct extends BaseModel {
 
   final String name;
   final double price;
-  final int quantity;
+  final double quantity;
 
   factory ReceiptProduct.fromJson(Map<String, dynamic> json) {
     return ReceiptProduct(
       name: json['product_name'] ?? '',
       price: (json['price'] ?? 0).toDouble(),
-      quantity: json['quantity'] ?? 1,
+      quantity: (json['quantity'] ?? 1).toDouble(),
     );
   }
